rename all component files
